\name{chi3class-class}
\Rdversion{1.0}
\docType{class}
\alias{chi3class}
\alias{chi3class-class}
\title{The class of common objects }
\description{
The class of objects common to the two kind of three-way indices.
}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{chi3class(...)}.
%%  ~~ describe objects here ~~ 
}
\section{Slots}{
  \describe{
    \item{\code{z}}{Object of class \code{"matrix"}. Terms of the partition of the three-way symmetrical and non symmetrical indices.}
  }
}
%\section{Extends}{
%Class \code{"\linkS4class{cacorporateplus}"}, directly.
%}
\section{Methods}{
No methods defined with this class  in the signature.
}
\references{
Beh EJ and Lombardo R (2014) Correspondence Analysis: Theory, Practice and New Strategies. John Wiley & Sons. \cr
Carlier A Kroonenberg PM (1996) Biplots and decompositions in two-way and three-way correspondence analysis. Psychometrika, 61, 355-373. \cr
Lancaster H O (1951) Complex contingency tables treated by the partition of the chi-square. Journal of Royal Statistical Society, Series B, 13, 242-249. \cr
Loisel S and Takane Y (2016) Partitions of Pearson's chi-square ststistic for frequency tables:  A comprehensive account. Computational Statistics, 31, 1429-1452.\cr
Lombardo R Carlier A D'Ambra L (1996)  Nonsymmetric correspondence analysis for three-way contingency tables. Methodologica,  4, 59-80. 
}
\author{
Lombardo R, Takane Y and Beh EJ}
